ref: PlayerDisplayのリファクタ
